:recycle: (components) refactor useResponsive and move to comp. lib
Two things in this PR. 1) refactored useResponsive - removed ResponsiveProvider context. It was necessary before as we
were relying on document to calculate widths. Whereas now we use a hook that react to window size updates. Hence we no
longer need to context provider; 2) moved the hook to the component library so it could be used in plugins and such.
