+++ packages/desktop-client/src/components/Page.tsx
@@ -1,12 +1,10 @@
 import React, { type ReactNode, type CSSProperties } from 'react';
 
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
+import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
-
-import { theme } from '../style';
-
-import { useResponsive } from './responsive/ResponsiveProvider';
 
 const HEADER_HEIGHT = 50;
