+++ packages/desktop-client/src/components/autocomplete/PayeeAutocomplete.test.tsx
@@ -8,7 +8,6 @@ import type { AccountEntity, PayeeEntity } from 'loot-core/types/models';
 import { AuthProvider } from '../../auth/AuthProvider';
 import { useCommonPayees } from '../../hooks/usePayees';
 import { TestProvider } from '../../redux/mock';
-import { ResponsiveProvider } from '../responsive/ResponsiveProvider';
 
 import {
   PayeeAutocomplete,
@@ -65,17 +64,15 @@ function renderPayeeAutocomplete(
   render(
     <TestProvider>
       <AuthProvider>
-        <ResponsiveProvider>
-          <div data-testid="autocomplete-test">
-            <PayeeAutocomplete
-              {...autocompleteProps}
-              onSelect={vi.fn()}
-              type="single"
-              value={null}
-              embedded={false}
-            />
-          </div>
-        </ResponsiveProvider>
+        <div data-testid="autocomplete-test">
+          <PayeeAutocomplete
+            {...autocompleteProps}
+            onSelect={vi.fn()}
+            type="single"
+            value={null}
+            embedded={false}
+          />
+        </div>
       </AuthProvider>
     </TestProvider>,
   );
