+++ packages/desktop-client/src/components/autocomplete/PayeeAutocomplete.tsx
@@ -14,8 +14,11 @@ import React, {
 import { Trans, useTranslation } from 'react-i18next';
 
 import { Button } from '@actual-app/components/button';
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
+import { SvgAdd, SvgBookmark } from '@actual-app/components/icons/v1';
 import { styles } from '@actual-app/components/styles';
 import { TextOneLine } from '@actual-app/components/text-one-line';
+import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 import { css, cx } from '@emotion/css';
 
@@ -28,10 +31,7 @@ import { type AccountEntity, type PayeeEntity } from 'loot-core/types/models';
 
 import { useAccounts } from '../../hooks/useAccounts';
 import { useCommonPayees, usePayees } from '../../hooks/usePayees';
-import { SvgAdd, SvgBookmark } from '../../icons/v1';
 import { useDispatch } from '../../redux';
-import { theme } from '../../style';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 
 import {
   Autocomplete,
