+++ packages/desktop-client/src/components/budget/BalanceWithCarryover.tsx
@@ -7,7 +7,10 @@ import React, {
 } from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
+import { SvgArrowThinRight } from '@actual-app/components/icons/v1';
 import { styles } from '@actual-app/components/styles';
+import { theme } from '@actual-app/components/theme';
 import { Tooltip } from '@actual-app/components/tooltip';
 import { View } from '@actual-app/components/view';
 import { css } from '@emotion/css';
@@ -15,9 +18,6 @@ import { css } from '@emotion/css';
 import { type TransObjectLiteral } from 'loot-core/types/util';
 
 import { useFeatureFlag } from '../../hooks/useFeatureFlag';
-import { SvgArrowThinRight } from '../../icons/v1';
-import { theme } from '../../style';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 import { type Binding } from '../spreadsheet';
 import { CellValue, CellValueText } from '../spreadsheet/CellValue';
 import { useFormat } from '../spreadsheet/useFormat';
