+++ packages/desktop-client/src/components/manager/ManagementApp.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { Text } from '@actual-app/components/text';
+import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 
 import { loggedIn } from 'loot-core/client/actions';
@@ -11,7 +13,6 @@ import { ProtectedRoute } from '../../auth/ProtectedRoute';
 import { Permissions } from '../../auth/types';
 import { useMetaThemeColor } from '../../hooks/useMetaThemeColor';
 import { useSelector, useDispatch } from '../../redux';
-import { theme } from '../../style';
 import { tokens } from '../../tokens';
 import {
   BackToFileListButton,
@@ -20,7 +21,6 @@ import {
 import { AppBackground } from '../AppBackground';
 import { LoggedInUser } from '../LoggedInUser';
 import { Notifications } from '../Notifications';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 import { useMultiuserEnabled, useServerVersion } from '../ServerContext';
 
 import { BudgetList } from './BudgetList';
