+++ packages/desktop-client/src/components/modals/ConfirmTransactionDeleteModal.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Button } from '@actual-app/components/button';
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { InitialFocus } from '@actual-app/components/initial-focus';
 import { Paragraph } from '@actual-app/components/paragraph';
 import { styles } from '@actual-app/components/styles';
@@ -10,7 +11,6 @@ import { View } from '@actual-app/components/view';
 import { type Modal as ModalType } from 'loot-core/client/modals/modalsSlice';
 
 import { Modal, ModalCloseButton, ModalHeader } from '../common/Modal';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 
 type ConfirmTransactionDeleteModalProps = Extract<
   ModalType,
