+++ packages/desktop-client/src/components/modals/CreateEncryptionKeyModal.tsx
@@ -4,10 +4,13 @@ import { Form } from 'react-aria-components';
 import { useTranslation, Trans } from 'react-i18next';
 
 import { ButtonWithLoading } from '@actual-app/components/button';
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { InitialFocus } from '@actual-app/components/initial-focus';
+import { Input } from '@actual-app/components/input';
 import { Paragraph } from '@actual-app/components/paragraph';
 import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
+import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 import { css } from '@emotion/css';
 
@@ -19,8 +22,6 @@ import { send } from 'loot-core/platform/client/fetch';
 import { getCreateKeyError } from 'loot-core/shared/errors';
 
 import { useDispatch } from '../../redux';
-import { theme } from '../../style';
-import { Input } from '../common/Input';
 import { Link } from '../common/Link';
 import {
   Modal,
@@ -28,7 +29,6 @@ import {
   ModalCloseButton,
   ModalHeader,
 } from '../common/Modal';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 
 type CreateEncryptionKeyModalProps = Extract<
   ModalType,
