+++ packages/desktop-client/src/components/modals/PayeeAutocompleteModal.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
+import { theme } from '@actual-app/components/theme';
+
 import { type Modal as ModalType } from 'loot-core/client/modals/modalsSlice';
 
 import { useAccounts } from '../../hooks/useAccounts';
 import { useNavigate } from '../../hooks/useNavigate';
 import { usePayees } from '../../hooks/usePayees';
-import { theme } from '../../style';
 import { PayeeAutocomplete } from '../autocomplete/PayeeAutocomplete';
 import {
   ModalCloseButton,
@@ -14,7 +16,6 @@ import {
   ModalTitle,
   ModalHeader,
 } from '../common/Modal';
-import { useResponsive } from '../responsive/ResponsiveProvider';
 
 type PayeeAutocompleteModalProps = Extract<
   ModalType,
