+++ packages/desktop-client/src/components/reports/graphs/tableGraph/ReportTableRow.tsx
@@ -5,7 +5,9 @@ import React, {
   type CSSProperties,
 } from 'react';
 
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { Text } from '@actual-app/components/text';
+import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 
 import {
@@ -22,8 +24,6 @@ import { type RuleConditionEntity } from 'loot-core/types/models/rule';
 import { useAccounts } from '../../../../hooks/useAccounts';
 import { useCategories } from '../../../../hooks/useCategories';
 import { useNavigate } from '../../../../hooks/useNavigate';
-import { theme } from '../../../../style';
-import { useResponsive } from '../../../responsive/ResponsiveProvider';
 import { Row, Cell } from '../../../table';
 import { showActivity } from '../showActivity';
