+++ packages/desktop-client/src/components/reports/reports/Spending.tsx
@@ -5,6 +5,7 @@ import { useParams } from 'react-router-dom';
 import { AlignedText } from '@actual-app/components/aligned-text';
 import { Block } from '@actual-app/components/block';
 import { Button } from '@actual-app/components/button';
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
 import { Paragraph } from '@actual-app/components/paragraph';
 import { Select } from '@actual-app/components/select';
 import { SpaceBetween } from '@actual-app/components/space-between';
@@ -33,7 +34,6 @@ import { FilterButton } from '../../filters/FiltersMenu';
 import { MobileBackButton } from '../../mobile/MobileBackButton';
 import { MobilePageHeader, Page, PageHeader } from '../../Page';
 import { PrivacyFilter } from '../../PrivacyFilter';
-import { useResponsive } from '../../responsive/ResponsiveProvider';
 import { SpendingGraph } from '../graphs/SpendingGraph';
 import { LegendItem } from '../LegendItem';
 import { LoadingIndicator } from '../LoadingIndicator';
