+++ packages/desktop-client/src/components/responsive/index.tsx
@@ -1,7 +1,8 @@
+import { useResponsive } from '@actual-app/components/hooks/useResponsive';
+
 import { LoadComponent } from '../util/LoadComponent';
 
 import type * as NarrowComponents from './narrow';
-import { useResponsive } from './ResponsiveProvider';
 import type * as WideComponents from './wide';
 
 const loadNarrow = () =>
